Cleanup of HipChatSession - moved into separate namespace and made logger part of constructor
